DAO implements Serializable
- This should make these into JavaBeans
- refactor Customer to User since customers are anonymous
